chore: re-organize a few files and update imports
